// src/app/model/FieldType.ts
export enum FieldType {
  SELECT = 2,
  TEXT_AREA = 3,
  DATA = 4,
  CPF = 5,
  MOEDA = 6,
  DATA_HORA = 7,
  CNPJ = 8,
  NUMERO = 10,
  MES_ANO = 11,
  ANO = 12,
  NUMERO_ORGAO = 13,
  NOME = 14,
  PHONE = 15,
  CELPHONE = 16,
  CEP = 17,
  SENHA = 18,
  CPF_CNPJ = 19,
  CHECKBOX = 20,
  MULTISELECT = 21,
  BOOLEAN = 22,
  AUTOCOMPLETE = 23,
  FILE = 24,
  DATE
}